+++ src/staticcore/java/theking530/api/heat/HeatTooltipUtilities.java
@@ -11,23 +11,23 @@ public class HeatTooltipUtilities {
 		return Component.literal(ChatFormatting.GRAY + "Heat Conductivity: ").append(GuiTextUtilities.formatConductivityToString(conductivity)).withStyle(ChatFormatting.AQUA);
 	}
 
-	public static MutableComponent getActiveTemperatureTooltip(int temperature) {
+	public static MutableComponent getActiveTemperatureTooltip(float temperature) {
 		return Component.literal(ChatFormatting.GRAY + "Temperature: ").append(GuiTextUtilities.formatHeatRateToString(temperature)).withStyle(ChatFormatting.GOLD);
 	}
 
-	public static MutableComponent getHeatGenerationTooltip(int heatGeneration) {
+	public static MutableComponent getHeatGenerationTooltip(float heatGeneration) {
 		return Component.literal(ChatFormatting.GRAY + "Heat Generation: ").append(GuiTextUtilities.formatHeatRateToString(heatGeneration)).withStyle(ChatFormatting.GREEN);
 	}
 
-	public static MutableComponent getOverheatingTooltip(int temperature) {
+	public static MutableComponent getOverheatingTooltip(float temperature) {
 		return Component.literal(ChatFormatting.GRAY + "Overheats At: ").append(GuiTextUtilities.formatHeatToString(temperature)).withStyle(ChatFormatting.YELLOW);
 	}	
 
-	public static MutableComponent getMaximumHeatTooltip(int capacity) {
+	public static MutableComponent getMaximumHeatTooltip(float capacity) {
 		return Component.literal(ChatFormatting.GRAY + "Maximum Capacity: ").append(GuiTextUtilities.formatHeatToString(capacity)).withStyle(ChatFormatting.RED);
 	}
 
-	public static MutableComponent getFreezingTooltip(int temperature) {
+	public static MutableComponent getFreezingTooltip(float temperature) {
 		return Component.literal(ChatFormatting.GRAY + "Freezes At: ").append(GuiTextUtilities.formatHeatToString(temperature)).withStyle(ChatFormatting.BLUE);
 	}
 }
